feat: Run a primary instance of MariaDB and replica
Add a Shipyard stack named "database-alpha" (so named because we expect
to run more than one database) which is a pair of MariaDB services,
configured for replication.

The database is primed with the (empty) databases and user accounts
described in the stack values.  Please update the shipyard role so that
this feature will work.

The stack values file is intentionally unencrypted because the values
are not sensitive.